Implementation of getLastId (which returns id from last insertion of data into user and entry) method, for UserDAO and EntryDAO

+++ src/main/java/com/srh/diet_tracker/EntryDAO.java
@@ -30,43 +30,61 @@ public class EntryDAO {
 
     // id wird vom entsprechenden Controller übergeben (last id für ControllerEntry, beliebige id für ControllerDayReview)
     public void updateEntryData(int id){
-        String url = "jdbc:sqlite:diet.db";
-        String sql = "UPDATE entry SET isSport = ? , "
-                + " calories = ? , "
-                + " sugar = ? "
-                + "WHERE id = ?";
+        if (id!=0) {
+            String url = "jdbc:sqlite:diet.db";
+            String sql = "UPDATE entry SET isSport = ? , "
+                    + " calories = ? , "
+                    + " sugar = ? "
+                    + "WHERE id = ?";
 
-        try (var conn = DriverManager.getConnection(url);
-             var pstmt = conn.prepareStatement(sql)){
-            pstmt.setBoolean(1,entry.isSport());
-            pstmt.setDouble(2,entry.getCalories());
-            pstmt.setDouble(3,entry.getSugar());
-            pstmt.setInt(4,id);
-            pstmt.executeUpdate();
+            try (var conn = DriverManager.getConnection(url);
+                 var pstmt = conn.prepareStatement(sql)) {
+                pstmt.setBoolean(1, entry.isSport());
+                pstmt.setDouble(2, entry.getCalories());
+                pstmt.setDouble(3, entry.getSugar());
+                pstmt.setInt(4, id);
+                pstmt.executeUpdate();
 
-        }catch (SQLException e){
-            System.err.println(e.getMessage());
+            } catch (SQLException e) {
+                System.err.println(e.getMessage());
+            }
+
         }
     }
 
     // id wird vom entsprechenden Controller übergeben (last id für ControllerUser, beliebige id für ControllerDayReview)
     public void deleteEntry (int id){
+        if (id!=0) {
+            String url = "jdbc:sqlite:diet.db";
+            String sql = "DELETE FROM entry WHERE id = ?";
+
+            try (var conn = DriverManager.getConnection(url);
+                 var pstmt = conn.prepareStatement(sql)) {
+                pstmt.setInt(1, id);
+                pstmt.executeUpdate();
+
+            } catch (SQLException e) {
+                System.err.println(e.getMessage());
+            }
+        }
+    }
+
+    public int getLastId(){
         String url = "jdbc:sqlite:diet.db";
-        String sql = "DELETE FROM entry WHERE id = ?";
+        String sql = "SELECT MAX(id) AS max_id FROM entry;";
 
         try (var conn = DriverManager.getConnection(url);
-             var pstmt = conn.prepareStatement(sql)) {
-            pstmt.setInt(1, id);
-            pstmt.executeUpdate();
+             var stmt = conn.createStatement();
+             var lastId= stmt.executeQuery(sql)) {
+            if (lastId.wasNull()) {
+                return lastId.getInt("max_id");
+            }
+            else {return 0;}
 
         } catch (SQLException e) {
             System.err.println(e.getMessage());
         }
+        return 0;
     }
-
-
-
-
-
 
 }
